Refactor applications into modular components
